product index除了检查是否存在于购物车外，其他完成

+++ www/static/app/modules/product_index.js
@@ -8,19 +8,22 @@
 angular.module(
     'xShowroom.product.index',
     [
-        'xShowroom.i18n', 'xShowroom.directives', 'ngSanitize'
+        'xShowroom.i18n', 'xShowroom.directives', 'ngSanitize', 'mgcrea.ngStrap'
     ]
 )
 .controller(
     'ProductIndexCtrl',
     [
-        '$scope', 'Product', 'Order',
-        function ($scope, Product, Order) {
+        '$scope', '$modal', 'Product', 'Order',
+        function ($scope, $modal, Product, Order) {
         	$scope.setProductCover = function (url) {
         		$scope.productCover = url;
         	};
         	
         	$scope.deleteProduct = function(productId, collectionId){
+        		if (!$scope.hasAuth){
+        			return;
+        		}
 				Product.destroy({
 					id: productId
 				}).success(function(res){
@@ -37,12 +40,38 @@ angular.module(
         			productionId: productId
         		}).success(function(res){
         			if (typeof(res) != 'object' || res.status) {
-     					alert('获取Cover数据失败，请检查！');
+        				$modal({title: 'Error Info', content: 'Add product to cart失败，请检查！'}).show();
+     					return;
+     				}
+        			$scope.isInCart = true;
+        		});
+        	};
+        	
+        	$scope.removeProductFromCart  = function(productId){
+        		Order.removeProductFromCart({
+        			productionId: productId
+        		}).success(function(res){
+        			if (typeof(res) != 'object' || res.status) {
+        				$modal({title: 'Error Info', content: 'remove product from cart失败，请检查！'}).show();
+     					return;
+     				}
+        			$scope.isInCart = false;
+        		});
+        	};
+        	
+        	$scope.checkProductIsExisted = function(productId){
+        		Order.checkProductIsExisted({
+        			productionId: $scope.productId
+        		}).success(function(res){
+        			if (typeof(res) != 'object' || res.status) {
+        				$modal({title: 'Error Info', content: '获取Cart数据失败，请检查！'}).show();
      					return;
      				}
         		});
         	};
         	
+        	
+        	$scope.checkProductIsExisted();
         }
     ]
 );
